Interstroke stratified analysis for predictor

Interstroke stratified analysis for predictor

Usage

is_figure_stratified_subgroups(
  interstroke_data,
  subgroups,
  model,
  limits = NULL,
  p_sig = NULL,
  breaks = NULL,
  text_size = NULL,
  drop_columns = NULL,
  include_overall = TRUE,
  long_title = TRUE,
  regression_type = NULL
)

Arguments

interstroke_data

Project specific Interstroke dataset.

subgroups

A list of subgroups for stratified analysis.

model

Model formula for analysis. Predictor should be first independant variable.

limits

**Optional** Limit the x axis in the format c(xmin, xmax).

p_sig

**Optional** Set the level of significance to hide non significant values. Default displays all p values <1.

breaks

**Optional** Set the sequence for breaks in the format seq(xmin, xmax, interval)

text_size

**Optional** Assign value for text size.

drop_columns

**Optional** Drop population and subgroup columns suitable for joint plots.

long_title

**Optional** Set to false to just display predictor as title

regression_type

either "unconditional" or "conditional". Conditional models require +strata(caseid) at end of model formula.

Value

A dataframe suitable for Interstroke forest ggplot function.

Examples